yapmate-leads/requirements.txt
# Apify SDK
apify-client==1.7.0

# OpenAI SDK
openai==1.12.0

# Google Sheets API
gspread==5.12.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Utilities
python-dotenv==1.0.1
schedule==1.2.1

# Web Scraping (for email extraction)
requests==2.31.0
beautifulsoup4==4.12.3

# Email Sending
resend==0.8.0
paramiko==3.4.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
